backend/requirements.txt
Flask>=2.3,<3.0
Flask-Cors>=3.0,<4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
